Fix overlay ruleset selector not inheriting from `OsuTabControl`
`ToolbarRulesetSelector` has a completely different way of playing sounds, so I'm intentionally not making it inherit `OsuTabControl` (it's practically a custom implementation of a tab control).
